LayoutTests/fast/profiler/start-and-stop-profiler-multiple-times-expected.txt

CONSOLE MESSAGE: line 59: Start the profiler the first time.
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  (program) start-and-stop-profiler-multiple-times.html 8
CONSOLE MESSAGE: line 75:   (program) start-and-stop-profiler-multiple-times.html 4
CONSOLE MESSAGE: line 75:   onload start-and-stop-profiler-multiple-times.html 23
CONSOLE MESSAGE: line 75:    startTest start-and-stop-profiler-multiple-times.html 10
CONSOLE MESSAGE: line 75:     (anonymous function) profiler-test-JS-resources.js 29
CONSOLE MESSAGE: line 75:      insertNewText profiler-test-JS-resources.js 17
CONSOLE MESSAGE: line 75:       createElement  0
CONSOLE MESSAGE: line 75:       createTextNode  0
CONSOLE MESSAGE: line 75:       appendChild  0
CONSOLE MESSAGE: line 75:       getElementById  0
CONSOLE MESSAGE: line 75:     endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
CONSOLE MESSAGE: line 59: Start the profiler the first time.
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  (program) start-and-stop-profiler-multiple-times.html 8
CONSOLE MESSAGE: line 75:   (program) start-and-stop-profiler-multiple-times.html 4
CONSOLE MESSAGE: line 75:   onload start-and-stop-profiler-multiple-times.html 23
CONSOLE MESSAGE: line 75:    startTest start-and-stop-profiler-multiple-times.html 10
CONSOLE MESSAGE: line 75:     (anonymous function) profiler-test-JS-resources.js 29
CONSOLE MESSAGE: line 75:      insertNewText profiler-test-JS-resources.js 17
CONSOLE MESSAGE: line 75:       createElement  0
CONSOLE MESSAGE: line 75:       createTextNode  0
CONSOLE MESSAGE: line 75:       appendChild  0
CONSOLE MESSAGE: line 75:       getElementById  0
CONSOLE MESSAGE: line 75:     endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
CONSOLE MESSAGE: line 59: Start the profiler the second time.
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  startTest start-and-stop-profiler-multiple-times.html 10
CONSOLE MESSAGE: line 75:   (anonymous function) profiler-test-JS-resources.js 29
CONSOLE MESSAGE: line 75:    insertNewText profiler-test-JS-resources.js 17
CONSOLE MESSAGE: line 75:     createElement  0
CONSOLE MESSAGE: line 75:     createTextNode  0
CONSOLE MESSAGE: line 75:     appendChild  0
CONSOLE MESSAGE: line 75:     getElementById  0
CONSOLE MESSAGE: line 75:   endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
CONSOLE MESSAGE: line 59: Start the profiler the first time.
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  (program) start-and-stop-profiler-multiple-times.html 8
CONSOLE MESSAGE: line 75:   (program) start-and-stop-profiler-multiple-times.html 4
CONSOLE MESSAGE: line 75:   onload start-and-stop-profiler-multiple-times.html 23
CONSOLE MESSAGE: line 75:    startTest start-and-stop-profiler-multiple-times.html 10
CONSOLE MESSAGE: line 75:     (anonymous function) profiler-test-JS-resources.js 29
CONSOLE MESSAGE: line 75:      insertNewText profiler-test-JS-resources.js 17
CONSOLE MESSAGE: line 75:       createElement  0
CONSOLE MESSAGE: line 75:       createTextNode  0
CONSOLE MESSAGE: line 75:       appendChild  0
CONSOLE MESSAGE: line 75:       getElementById  0
CONSOLE MESSAGE: line 75:     endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
CONSOLE MESSAGE: line 59: Start the profiler the second time.
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  startTest start-and-stop-profiler-multiple-times.html 10
CONSOLE MESSAGE: line 75:   (anonymous function) profiler-test-JS-resources.js 29
CONSOLE MESSAGE: line 75:    insertNewText profiler-test-JS-resources.js 17
CONSOLE MESSAGE: line 75:     createElement  0
CONSOLE MESSAGE: line 75:     createTextNode  0
CONSOLE MESSAGE: line 75:     appendChild  0
CONSOLE MESSAGE: line 75:     getElementById  0
CONSOLE MESSAGE: line 75:   endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
CONSOLE MESSAGE: line 59: Start the profiler the third time.
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  startTest start-and-stop-profiler-multiple-times.html 10
CONSOLE MESSAGE: line 75:   (anonymous function) profiler-test-JS-resources.js 29
CONSOLE MESSAGE: line 75:    insertNewText profiler-test-JS-resources.js 17
CONSOLE MESSAGE: line 75:     createElement  0
CONSOLE MESSAGE: line 75:     createTextNode  0
CONSOLE MESSAGE: line 75:     appendChild  0
CONSOLE MESSAGE: line 75:     getElementById  0
CONSOLE MESSAGE: line 75:   endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
This page's JavaScript calls console.profile() and console.profileEnd() three times. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. In the profiler there should be three profiles.
This is inserted Text

This is inserted Text

This is inserted Text
